add component MegaStats to MegaPokemon
