+++ src/component/MegaPokemon.jsx
@@ -1,12 +1,9 @@
 import React, { useState } from "react";
-import { Swiper, SwiperSlide } from "swiper/react";
-import { Navigation } from "swiper/modules";
-import "swiper/css";
-import "swiper/css/navigation";
 import MegaTabs from "./megaDetails/MegaTabs";
 import MegaAbilities from "./megaDetails/MegaAbilities";
 import MegaHeader from "./megaDetails/MegaHeader";
 import MegaSwiper from "./megaDetails/MegaSwiper";
+import MegaStats from "./megaDetails/MegaStats";
 
 const MegaEvolutions = ({ megaEvolutions }) => {
     const [openIndex, setOpenIndex] = useState(null);
@@ -19,23 +16,6 @@ const MegaEvolutions = ({ megaEvolutions }) => {
     const toggleAccordion = (key) => {
         setOpenIndex(openIndex === key ? null : key);
     }
-
-    const getStatColor = (value) => {
-        if (value >= 100) return "bg-green-500"
-        if (value > 80) return "bg-yellow-500"
-        if (value > 40) return "bg-orange-500"
-        if (value > 20) return "bg-yellow-500"
-        return "bg-red-500"
-    }
-      
-    const statMapping = {
-        'hp': 'HP',
-        'attack': 'Atk',
-        'defense': 'Def',
-        'special-attack': 'SpAtk',
-        'special-defense': 'SpDef',
-        'speed': 'Spd',
-    };
 
     return (
             <div className="">
@@ -62,25 +42,7 @@ const MegaEvolutions = ({ megaEvolutions }) => {
                                     </>
                                 ) : (
                                     <>
-                                        <div className="space-y-4 pt-4 w-full">
-                                            <h2 className="text-xl font-semibold mb-2">Estadísticas</h2>
-                                            {mega.stats?.map((stat) => (
-                                                <div key={stat.name} className="space-y-1 w-full">
-                                                    <div className="flex justify-between text-sm w-full">
-                                                        <span className="text-sm font-medium">
-                                                            {statMapping?.[stat.name] || stat.name}
-                                                        </span>
-                                                        <span className="text-sm font-medium">{stat.base}</span>
-                                                    </div>
-                                                    <div className="h-2 w-full bg-gray-200 rounded-full overflow-hidden">
-                                                        <div 
-                                                        className={`h-2 transition-all duration-500 ${getStatColor(stat.base)}`} 
-                                                        style={{ width: `${(stat.base / 150) * 100}%` }}
-                                                        ></div>
-                                                    </div>
-                                                </div>
-                                            ))}
-                                        </div>
+                                        <MegaStats stats={mega.stats} />
                                     </>
                                 )}
                             </div>
